rename splash screen tsx
