+++ app-design-v2/src/components/SplashModal.tsx
